requirements.txt
Django
djangorestframework
dice
psycopg2-binary
coverage
python_coveralls
pygame
requests
drf-nested-routers
django-gm2m
